+++ sharkfuser/include/fusilli/attributes/attributes.h
@@ -19,7 +19,7 @@
 #define FUSILLI_ATTRIBUTES_ATTRIBUTES_H
 
 #include "fusilli/attributes/tensor_attributes.h"
-#include "fusilli/context.h"
+#include "fusilli/graph/context.h"
 #include "fusilli/types.h"
 
 #include <memory>
