+++ sharkfuser/include/fusilli/backend/backend.h
@@ -11,8 +11,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef FUSILLI_BACKEND_H
-#define FUSILLI_BACKEND_H
+#ifndef FUSILLI_BACKEND_BACKEND_H
+#define FUSILLI_BACKEND_BACKEND_H
 
 #include <string>
 #include <unordered_map>
@@ -47,4 +47,4 @@ static const std::unordered_map<Backend, std::vector<std::string>>
 
 } // namespace fusilli
 
-#endif // FUSILLI_BACKEND_H
+#endif // FUSILLI_BACKEND_BACKEND_H
