+++ sharkfuser/include/fusilli/graph/context.h
@@ -12,10 +12,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef FUSILLI_CONTEXT_H
-#define FUSILLI_CONTEXT_H
+#ifndef FUSILLI_GRAPH_CONTEXT_H
+#define FUSILLI_GRAPH_CONTEXT_H
 
-#include "fusilli/backend.h"
+#include "fusilli/backend/backend.h"
 #include "fusilli/types.h"
 
 #include <string>
@@ -71,4 +71,4 @@ private:
 
 } // namespace fusilli
 
-#endif // FUSILLI_CONTEXT_H
+#endif // FUSILLI_GRAPH_CONTEXT_H
