+++ sharkfuser/include/fusilli/node/node.h
@@ -14,7 +14,7 @@
 #ifndef FUSILLI_NODE_NODE_H
 #define FUSILLI_NODE_NODE_H
 
-#include "fusilli/context.h"
+#include "fusilli/graph/context.h"
 #include "fusilli/logging.h"
 
 #include <memory>
